requirements.txt
imgcat >= 0.5.0
Pillow >= 7.2.0